Fix queries
Previously we queried only for packages with outstanding revenue, and of
those we queued redistribution and tagged them as noComp. This meant
that packages could exist in the DB that are noComp in `meta` but not
marked as such. We prefer to keep the package document as consistent as
possible, so this commit finds no comp packages with outstanding revenue
OR without the noComp flag. Any revenue on the returned packages is
queued for redistribution, and all the packages are marked as noComp in
the end.

As an additional optimization, I filter out revenue-less packages before
determining which packages have no deps. The no-dep check is only
important when redistributing revenue, so it can be skipped for packages
where decompress only plans on flipping the flag.